My interpretation of Conways Game of Life
Added my project for my interpretation of Conway's Game of Life. This includes the exe of my final release after I finished the code. This my first big project. The code requires the wxWidgets Binaries to work which is an external file that must be download and dropped into the same same folder as the Game of Life.sln file.